pyproject.toml
[project]
name = "image-analyser"
version = "0.1.0"
description = "Determines which images are similar"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]
